requirements.txt
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.40.2
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.1
gevent==25.5.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.10.1
python-decouple==3.8
python-dotenv==1.1.1
python-engineio==4.11.2
python-nmap==0.7.1
python-socketio==5.12.1
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
